Part 5 completed.
Finished Part 5 which involved more progress in connecting the database to frontend.
- Can create new lists
- Can select new lists and show that list as the active selected list.
- Can display the selected lists tasks

I also fixed a bug that had me redo the entirety of Part's 2-5 again in a new GitHub Repository because of a error that I initially thought was my own doing. I thought I changes something that I simply could not find or remember which was causing errors. The problem I was experiencing gets fixed in Part 10 at ~1:15. The if else statement fixed my error. Back on to Part 6 now.